Update 000002.php
Your code is attempting to add a column named "language" to the user table and set the language value of all rows to "en". However, the same issue arises again: SQLite does not support the use of the pragma_table_info function in a SELECT query. You can use the PRAGMA table_info(table_name) statement to get the table information, and then handle the returned results in PHP.